refactor(eslint-plugin-dialtone, dialtone-css): DLT-3329 extract shared class-attribute autofix helper
Both `deprecated-pixel-utility-classes` and `deprecated-radius-utility-classes`
had ~17 lines of identical scaffolding (token-boundary anchors, combined
DETECT regex, VAttribute visitor, fix function with quote preservation). The
recent `\b` → `(?<=^|\s)` fix had to land in both places — proof of drift
risk. Extract the common parts into `lib/util/class-attribute-rule.js`.

Also trims internal-reference noise from `dialtone-generators.cjs` comments.

Net: -45 lines across the two rules.
